Split out spec/support
