+++ spec/spec_helper.rb
@@ -9,35 +9,4 @@ require 'webmock/rspec'
 ENV['CODECLIMATE_REPO_TOKEN'] = "172754c1bf9a3c698f7770b9fb648f1ebb214425120022d0b2ffc65b97dff531"
 ENV['CODECLIMATE_API_HOST']   = "http://cc.dev"
 
-module TestHelper
-  def inflate(string)
-    reader = Zlib::GzipReader.new(StringIO.new(string))
-    reader.read
-  end
-
-  def capture_requests(stub)
-    requests = []
-    stub.to_return { |r| requests << r; {body: "hello"} }
-    requests
-  end
-
-  def capture_io
-    stdout = $stdout
-    stderr = $stderr
-    $stdout = StringIO.new
-    $stderr = StringIO.new
-
-    yield if block_given?
-
-    [$stdout, $stderr]
-  ensure
-    $stdout = stdout
-    $stderr = stderr
-  end
-end
-
-RSpec.configure do |c|
-  c.include TestHelper
-end
-
-
+Dir.glob("spec/support/**/*.rb").sort.each(&method(:load))
